// 25032026/sce-front-auditoria/sce-front-auditoria/src/app/model/actaPorCorregirBean.ts
import {ActasPorCorregirListBean} from "./actasPorCorregirListBean";
import {AgrupolPorCorregirBean} from "./agrupolPorCorregirBean";
import {ItemPorCorregirBean} from "./itemPorCorregirBean";

export class ActaPorCorregirBean{
  acta: ActasPorCorregirListBean;
  agrupacionesPoliticas: Array<AgrupolPorCorregirBean>;
  cvas: ItemPorCorregirBean;
  observaciones: Array<ItemPorCorregirBean>;

}